moved more components to mips-lib
