+++ mips-lib/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+MIPS_disassembly = "0.1.2"
 anyhow = "1.0.72"
 clap = { version = "4.3.15", features = ["derive"] }
 elf = "0.7.4"
