+++ mips-lib/src/components/mips_branch_logic.rs
@@ -1,13 +1,13 @@
 // use std::fmt::Alignment;
-#[cfg(feature = "gui-egui")]
-use crate::common::EguiComponent;
-use crate::common::{
-    Component, Condition, Id, Input, InputPort, OutputType, Ports, SignalValue, Simulator,
-};
 use log::*;
 use serde::{Deserialize, Serialize};
 use std::any::Any;
 use std::rc::Rc;
+#[cfg(feature = "gui-egui")]
+use syncrim::common::EguiComponent;
+use syncrim::common::{
+    Component, Condition, Id, Input, InputPort, OutputType, Ports, SignalValue, Simulator,
+};
 
 pub const BRANCH_OP_ID: &str = "branch_op_in";
 pub const BRANCH_RT_ID: &str = "branch_rt_in";
