+++ mips-lib/src/gui_egui/components/full_adder.rs
