+++ mips-lib/src/gui_egui/components/mips_control_unit.rs
