+++ mips-lib/src/gui_egui/components/mod.rs
@@ -1,0 +1,9 @@
+mod full_adder;
+mod mips_branch_logic;
+mod mips_control_unit;
+mod mips_dm;
+mod mips_im;
+mod mips_instruction_splitter;
+mod mips_jump_merge;
+mod mips_reg_file;
+mod sz_extend;
