+++ src/components/mod.rs
@@ -4,16 +4,7 @@ mod clk;
 mod constant;
 mod cross;
 mod equals;
-//mod full_adder;
 mod mem;
-mod mips_branch_logic;
-//mod mips_controll_unit;
-//mod mips_dm;
-mod mips_im;
-mod mips_instruction_splitter;
-mod mips_jump_merge;
-pub mod mips_mem_struct;
-mod mips_reg_file;
 mod mux;
 mod probe;
 mod probe_assert;
@@ -33,16 +24,7 @@ pub use clk::*;
 pub use constant::*;
 pub use cross::*;
 pub use equals::*;
-pub use mips_mem_struct::*;
-//pub use full_adder::*;
 pub use mem::*;
-pub use mips_branch_logic::*;
-//pub use mips_controll_unit::*;
-//pub use mips_dm::*;
-pub use mips_im::*;
-pub use mips_instruction_splitter::*;
-pub use mips_jump_merge::*;
-pub use mips_reg_file::*;
 pub use mux::*;
 pub use probe::*;
 pub use probe_assert::*;
