--- src/gui_egui/components/mips_jump_merge.rs
@@ -1,63 +1,0 @@
-use crate::common::{EguiComponent, Id, Ports, Simulator};
-use crate::components::JumpMerge;
-use crate::gui_egui::editor::{EditorMode, EditorRenderReturn, GridOptions};
-use crate::gui_egui::gui::EguiExtra;
-use crate::gui_egui::helper::basic_component_gui;
-use egui::{Rect, Response, RichText, Ui, Vec2};
-
-#[typetag::serde]
-impl EguiComponent for JumpMerge {
-    fn render(
-        &self,
-        ui: &mut Ui,
-        _context: &mut EguiExtra,
-        simulator: Option<&mut Simulator>,
-        offset: Vec2,
-        scale: f32,
-        clip_rect: Rect,
-        _editor_mode: EditorMode,
-    ) -> Option<Vec<Response>> {
-        basic_component_gui(self, &simulator, ui.ctx(), offset, scale, clip_rect, |ui| {
-            ui.label(RichText::new("Jump Merge").size(12f32 * scale));
-        })
-    }
-
-    fn render_editor(
-        &mut self,
-        ui: &mut egui::Ui,
-        context: &mut EguiExtra,
-        simulator: Option<&mut Simulator>,
-        offset: egui::Vec2,
-        scale: f32,
-        clip_rect: egui::Rect,
-        _id_ports: &[(Id, Ports)],
-        _grid: &GridOptions,
-        editor_mode: EditorMode,
-    ) -> EditorRenderReturn {
-        self.render(
-            ui,
-            context,
-            simulator,
-            offset,
-            scale,
-            clip_rect,
-            editor_mode,
-        );
-        EditorRenderReturn {
-            delete: false,
-            resp: None,
-        }
-    }
-
-    fn set_pos(&mut self, pos: (f32, f32)) {
-        self.pos = pos;
-    }
-
-    fn get_pos(&self) -> (f32, f32) {
-        self.pos
-    }
-
-    fn top_padding(&self) -> f32 {
-        20f32
-    }
-}
