+++ src/gui_egui/components/mod.rs
@@ -4,15 +4,7 @@ mod clk;
 mod constant;
 mod cross;
 mod equal;
-//mod full_adder;
 mod mem;
-mod mips_branch_logic;
-//mod mips_controll_unit;
-//mod mips_dm;
-mod mips_im;
-mod mips_instruction_splitter;
-mod mips_jump_merge;
-mod mips_reg_file;
 mod mux;
 mod probe;
 mod probe_assert;
@@ -22,6 +14,5 @@ mod probe_stim;
 mod register;
 mod sext;
 mod shift_left_const;
-//mod sz_extend;
 mod wire;
 mod zero_extend;
