+++ src/gui_egui/mod.rs
@@ -6,7 +6,6 @@ pub mod helper;
 mod keymap;
 mod library;
 mod menu;
-pub mod mips_mem_view_window;
 
 #[cfg(feature = "components")]
 pub mod components;
